# model-pricing-backend/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "大模型中转平台：模型数据中心 + 进货/出货两条定价线。统一响应包 {code,message,data,requestId}；金额一律字符串传输；时间 ISO 8601 带时区。\n错误码：400 参数校验、401 未认证、403 无权限、409 幂等/状态冲突、423 主体冻结、429 限流。\n写操作（产生业务新版本/影响资金合同）必须携带 Idempotency-Key 头。\n分页：?page=1\u0026size=20 或 ?cursor=\u0026size=，统一返回 {list,total,page,size}。",
        "title": "模型管理与定价中心 API",
        "contact": {
            "name": "后端"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/api/internal/change-requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "官方价变更单与价目表发布进度统一查询。type 取值：DEPRECATE / PRICE_UP / PRICE_DOWN / PRICE_BOOK_PUBLISH / PRICE_BOOK_ROLLBACK / SPECIAL_PRICE；status 取值：PENDING / APPROVED / REJECTED。返回每单的 steps_total / steps_approved / pending_step_no / pending_role。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "变更单"
                ],
                "summary": "变更单列表（含审批进度摘要）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "变更类型（空=全部）",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态（空=全部）",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "SKU ID（空=全部）",
                        "name": "sku_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码（默认 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小（默认 20，上限 200）",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code=0；data={list,total,page,size}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "400": {
                        "description": "code=10001 参数非法",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 未认证",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/internal/change-requests/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "前端按 change_request_id 恢复审批状态（价目表发布、官方价变更均适用），不再依赖浏览器缓存。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "变更单"
                ],
                "summary": "变更单详情（含审批步骤链）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "change_request.id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code=0；data=ChangeRequestDetail（含 steps[]）",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "400": {
                        "description": "code=10001 参数非法",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 未认证",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "code=10004 变更单不存在",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务错误码，0=成功",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "业务数据",
                    "type": "object"
                },
                "message": {
                    "description": "错误消息，成功时 \"ok\"",
                    "type": "string",
                    "example": "ok"
                },
                "requestId": {
                    "description": "请求追踪号（X-Request-Id）",
                    "type": "string",
                    "format": "uuid",
                    "example": "01929d81-..."
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer \u003ctoken\u003e，登录后获得",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}